+++ src/main/java/org/superbiz/struts/UserServiceImpl.java
@@ -34,7 +34,7 @@ public class UserServiceImpl implements UserService {
         manager.persist(user);
     }
 
-    public User find(int id) {
+    public User find(long id) {
         return manager.find(User.class, id);
     }
